requirements.txt
django==6.0.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
openpyxl==3.1.2
psycopg2-binary==2.9.11
pgvector==0.3.6
google-generativeai==0.8.6
PyMuPDF==1.27.2
python-docx==1.2.0
mammoth==1.12.0
numpy==2.4.3
pillow==12.1.1
requests==2.32.5
tqdm==4.67.3
whitenoise==6.9.0
